backend/requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==2.0.1  # Keeping this version for now
Flask-Cors==5.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.3
Werkzeug==2.0.1  # Keeping this version for now
twelvedata[pandas]==1.2.24
sqlalchemy==2.0.36
sec-api==1.0.25
flask-jwt-extended==4.7.1
bcrypt==4.3.0
matplotlib==3.10.1